Move test_print_new_cast_table to test_casting_unittests.py and update imports
